Extract `flatObject` to own file
